Fix #17: Promote histogram bucket counters from int32 to int64
int32 overflows at ~2.1 billion, reachable in ~2.7 hours at 220K
events/sec hitting the same bucket.  After wrap the counter goes
negative -- silent data corruption.

Change histogram[] from int32 to int64 in WaitEventTimingEntry,
update pg_proc.dat type from _int4 to _int8, and update both
construct_array_builtin() calls to use Int64GetDatum/INT8OID.

Memory impact: +64 bytes per entry, ~204 KB per backend, ~19.9 MB
at 100 backends.

+++ src/include/utils/wait_event_timing.h
@@ -86,7 +86,7 @@ typedef struct WaitEventTimingEntry
 	int64		count;			/* number of occurrences */
 	int64		total_ns;		/* total wait duration in nanoseconds */
 	int64		max_ns;			/* longest single wait in nanoseconds */
-	int32		histogram[WAIT_EVENT_TIMING_HISTOGRAM_BUCKETS];
+	int64		histogram[WAIT_EVENT_TIMING_HISTOGRAM_BUCKETS];
 } WaitEventTimingEntry;
 
 /*
